[WebAssembly] Initial support for experimental big endian variant.
Adds support for a new "wasm32be" architecture to LLVM, Clang and LLD.
This is an experimental version of WebAssembly, with the main difference
being that all memory accesses are defined to be big endian, rather than
little endian.
The object file header version is set to 0xFFFFFFFF (-1) to specify this
is an incompatible variant.